+++ app/src/main/java/kz/kazpost/loadingarea/ui/scan/ScanRepository.kt
@@ -14,7 +14,7 @@ interface ScanRepository {
     fun rememberAddedParcel(shpi: String, tInvoiceNumber: String)
     fun verifyThatAllParcelsAreIncluded(
         factParcels: List<String>,
-        tInvoiceId: Int,
+        tInvoiceId: Long,
         index: Int,
         tInvoiceNumber: String
     ): Flow<Response<MissingShpisModel>>
